pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.3",
    "django>=5.2",
    "django-tailwind>=4.0.1",
    "google-genai>=1.18.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
